Render case studies in government-frontend
We have been publishing case studies to publishing-api for some time, and
government-frontend is now ready to handle case studies we can finally use it
to render them in production.

The migration is required to republish all existing case studies to
publishing-api with the updated rendering_app. publishing-api will send this on
to router-api/router so that it routes requests for the URLs to
government-frontend instead of whitehall-frontend.